KVM: X86: Add kvm_x86_ops to get the max page level for the TDP
When running an SEV-SNP VM, the sPA used to index the RMP entry is
obtained through the TDP transalation (gva->gpa->spa). The TDP page
level is checked against the page level programmed in the RMP entry.
If the page level does not match, then it will cause a nested page
table fault -- with the RMP bit set to indicate the RMP violation.

To resolve the fault, we must match the page levels between the TDP
and RMP entry. Add a new kvm_x86_op (get_tdp_max_page_level) that
can be used to query the existing the RMP page size. The page fault
handler will call the architecture code to get the maximum allowed page
level for the GPA and limit the TDP page level to it.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>