KVM: SVM: add support to handle MSR based Page State Change VMGEXIT
SEV-SNP VMs can ask hypervisor to change the page state in the RMP table
to be private or shared using the Page State Change MSR protocol as
defined in the GHCB specification section 2.5.1.

Before changing the page state in the RMP entry, we lookup the page in
the TDP to make sure that there is a valid mapping for it. If the mapping
exist then try to find a workable page level between the TDP and RMP for
the page. If the page is not mapped in the TDP, then create a fault such
that it gets mapped before we change the page state in the RMP entry.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>