mm: add support to split the large THP based on RMP violation
When SEV-SNP is enabled globally in the system, an RMP violation will
raise a page-fault exception. If the violation was seen for the
userspace memory range then we could resolve the fault by spliting the
backing page to lower level.

The architecture specific code will read the RMP entry to determine
if the fault can be resolved by spliting the backing page. And propagate
the request to split the page by setting newly intrdouced fault flag
(FAULT_FLAG_PAGE_SPLIT). If the fault can not be resolved by split,
then send a SIGBUS signal to terminate the process.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>