KVM: X86: Introduce kvm_mmu_get_tdp_walk() for SEV-SNP use
The SEV-SNP VMs may call the page state change VMGEXIT to add the gpa
as private or shared in the RMP table. The page state change VMGEXIT
will contain the RMP page level to be used in the RMP entry. If the
page level between the TDP and RMP does not match then it will result
in nested-page-fault (RMP violation).

The kvm_mmu_get_tdp_walk() will be used by the SEV-SNP VMGEXIT handler
to get the current page-level in the TDP for the given GPA. And calculate
a workable page level. e.g If a gpa is mapped as 4K-page in the TDP, but
the guest requested to add the gpa as a 2M in the RMP entry. The page
state change VMGEXIT handler will break the 2M request into multiple of
4K-pages to keep the RMP and TDP page-levels in sync.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>