X86: kernel: make the bss.decrypted section shared in RMP table
The encryption attribute for the bss.decrypted region is cleared in the
initial page table build. This is because the section contains the data
that need to be shared between the guest and the hypervisor.

When SEV-SNP is active, just clearing the encryption attribute in the
page table is not enough. We also need to make the page shared in the
RMP table.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>