[INTERNAL] Print S5_RESET_STATUS at boot
The following register contains bits that indicate the cause for the
previous reset.

	PMx000000C0 (FCH::PM::S5_RESET_STATUS)

This is useful for debug, etc., so print it during boot.

The register is accessed indirectly through a "PM" port in the FCH. The
data needs to be read a byte at a time. Only the primary FCH needs to be
read, so do this on the BSP.

Signed-off-by: Yazen Ghannam <yazen.ghannam@amd.com>