WORKAROUND?: KVM: allow encryption bit to be set for CR3 for SEV hosts
The current logic relies on CPUID values supplied by userspace to
determine the valid address bits, but in the case of kselftests these
registers are set as soon as a VCPU is instantiated, so we never get
the chance. As a result, attempting to set the encryption gets flagged
as invalid when we call KVM_SET_SREGS. Fix this by always allowing the
encryption bits allowed by the host.

Signed-off-by: Michael Roth <michael.roth@amd.com>