KVM: selftests: move vm_phy_pages_alloc() earlier in file
Subsequent patches will break some of this code out into file-local
helper functions, which will be used by functions like vm_vaddr_alloc(),
which currently are defined earlier in the file, so a forward
declaration would be needed.

Instead, move it earlier in the file, just above vm_vaddr_alloc() and
and friends, which are the main users.

Signed-off-by: Michael Roth <michael.roth@amd.com>