[WIP] kvm: x86/cpuid: Fix Architectural Performance Monitoring support
CPUID 0xA provides information on Architectural Performance
Monitoring features. It advertises a PMU version which QEMU
uses to determine the availability of additional global
control and status MSRs to manage the PMCs.

Upon receiving a KVM_GET_SUPPORTED_CPUID ioctl request for
the same, the kernel constructs return values based on the
x86 PMU capabilities irrespective of the vendor.

The CPUID function and the additional MSRs are not supported
on AMD processors. This breaks guest startup on upcoming AMD
processors for which the PMU version is set to 2. Hence, the
values returned should be zeros.

Fixes: a6c06ed1a60a ("KVM: Expose the architectural performance monitoring CPUID leaf")
Reported-by: Vasant Hegde <vasant.hegde@amd.com>
Signed-off-by: Sandipan Das <sandipan.das@amd.com>