gfs2: Add GL_NOPID flag for process-independent glock holders
Add a GL_NOPID flag to indicate that a holder isn't associated with the
current process.  This is useful for iopen glocks which are associated
with open files, as well as journal glock holders and similar which are
associated with the filesystem.

When the GL_NOPID flag is set, the pid and process name are only
reported while that process is still waiting for the holder to be
acquired.  After that, the pid is reported as 0 and the process name is
reported as "(none)".

When this flag is used (see the next patch), processes will no longer be
reported as holding glocks which they are not actually holding in the
glocks dump file.  So when a process is reported as having "(ended)",
this will actually indicate a bug.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>