KVM: SVM: Add support to handle GHCB GPA register VMGEXIT
SEV-SNP guests are required to perform a GHCB GPA registration. Before
using a GHCB GPA for a vCPU the first time, a guest must register the
vCPU GHCB GPA. If hypervisor can work with the guest requested GPA then
it must respond back with the same GPA otherwise return -1.

On VMEXIT, Verify that GHCB GPA matches with the registered value. If a
mismatch is detected then abort the guest.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
[mdr: update for restrictedmem, where page-state changes to GHCB are no
      longer a risk to kernel accessors, since the shared/private
      versions of a guest page will necessarily be backed by 2 separate
      PFNs.]
Signed-off-by: Michael Roth <michael.roth@amd.com>