KVM: SVM: Introduce ops for the post gfn map and unmap
When SEV-SNP is enabled in the guest VM, the guest memory pages can
either be a private or shared. A write from the hypervisor goes through
the RMP checks. If hardware sees that hypervisor is attempting to write
to a guest private page, then it triggers an RMP violation #PF.

To avoid the RMP violation with GHCB pages added new post_{map,unmap}_gfn
functions to verify if its safe to map GHCB pages.  Need to add generic
post_{map,unmap}_gfn() ops that can be used to verify that its safe
to map a given guest page in the hypervisor. Uses a spinlock to protect
against the page state change for existing mapped pages.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off by: Ashish Kalra <ashish.kalra@amd.com>