[WIP] perf/x86/amd/lbr: Add LbrExtV2 branch record support
If AMD Last Branch Record Extension Version 2 (LbrExtV2) is
detected, enable it alongside the freeze feature when an
event requests branch stack i.e. PERF_SAMPLE_BRANCH_STACK.

Each branch record is represented by a pair of registers,
LBR From and LBR To. The freeze feature prevents any updates
to these registers once a PMC overflows. The contents remain
unchanged until the freeze bit is cleared by the PMI handler.

The branch records are read and copied to sample data before
unfreezing. However, only valid entries, i.e. those denoting
branches with either the Valid bit or the Spec bit set, are
copied. There is no additional register to denote which of
these register pairs represent the top of the stack (TOS) as
internal register renaming ensures that the first pair is
representing the most recent branch and so on.

The LBR registers are per-thread resources and are cleared
explicitly whenever a new task is scheduled in. There are no
special implications on the contents of these registers when
transitioning to deep C-states.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>