[WIP] perf/x86/amd/lbr: Add LbrExtV2 hardware branch filter support
If AMD Last Branch Record Extension Version 2 (LbrExtV2) is
detected, enable it when an event requests branch stack and
also apply the requested branch filter.

The LBR Branch Select register has bits which if set, can
suppress recording of the following branch types:
  * CPL = 0 (Kernel only)
  * CPL > 0 (Userspace only)
  * Conditional Branches
  * Near Relative Calls
  * Near Indirect Calls
  * Near Returns
  * Near Indirect Jumps (excluding Near Indirect Calls and Near Returns)
  * Near Relative Jumps (excluding Near Relative Calls)
  * Far Branches

Bit combinations are mapped to PERF_SAMPLE_BRANCH_* and the
filter is applied whenever the PMU is enabled with an event
that requested branch stack.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>