[WIP] perf/x86/amd/lbr: Add LbrExtV2 software branch filter support
With AMD Last Branch Record Extension Version 2 (LbrExtV2),
it is necessary to process the branch records further as the
hardware filter cannot distinguish between certain types of
branches. E.g. to capture system calls, it is necessary to
enable the far branch filter but that will also capture jump
and return instructions. Use the branch classifier to remove
irrelevant records in such cases.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>