KVM: SEV: Handle KVM_HC_MAP_GPA_RANGE hypercall
KVM_HC_MAP_GPA_RANGE hypercall is used by the SEV guest to notify a
change in the page encryption status to the hypervisor. Add support
to maintain the page status in the shared bitmap in arch memslot.

The hypercall exits to userspace with KVM_EXIT_HYPERCALL exit code,
currently this is used for explicit memory conversion between
shared/private for memfd based private memory.

Signed-off-by: Nikunj A Dadhania <nikunj@amd.com>
[mdr: implement via new update_gpa_range x86 op]
Signed-off-by: Michael Roth <michael.roth@amd.com>