*mm: shmem: include PFN in invalidate/fallocate notifications
Don't quite recall why this was needed. Likely because for priv->shared
invalidation, notifications are issues with the pages locked via
find_lock_entries(), so memfile_get_lock_pfn() to retrieve them causes
deadlock. Might not be necessary for fallocate notifications.

A non-locking memfile_get_pfn() might be an alternative, need to follow
up.

Because of this notifications need to be issues for contiguous ranges,
since the PFN is treating as the starting PFN/index.

Also add a flag on the KVM side to differentiate invalidate from
fallocate notifications since things like RMP table updates for SEV-SNP
will differ between the two.

Signed-off-by: Michael Roth <michael.roth@amd.com>