*KVM: include encryption bits in kvm_page_fault flags
SEV-SNP will set bit 34 in page flags to indicate whether an NPT fault
was for an encrypted page or not. kvm_page_fault only tracks flags for
lower 32 bits currently.

The established way to handle this seems to be to record the bit as a
separate flag in the struct, but for some flexibility for testing just
store the whole 64-bits as well, along with the RMP bit (31).

Signed-off-by: Michael Roth <michael.roth@amd.com>