*KVM: SEV: add RMP #NPF handling for private memory
Page state changes are handled by userspace now, so if an RMP fault is
triggered as a result of a RMP NPT fault, exit to userspace just like
with explicit conversions.

RMP NPT faults can also occur if the guest pvalidates a 2M page as 4K,
in which case the RMP entries need to be PSMASH'd. Handle this case
immediately in the kernel.

Signed-off-by: Michael Roth <michael.roth@amd.com>